+++ src/components/Navbar/Navbar.js
@@ -1,23 +1,39 @@
-import { AppBar, Box, Grid, Toolbar } from "@mui/material";
-import { NavLink, Link } from "react-router-dom";
+import { AppBar, Box, Toolbar } from "@mui/material";
+import { useEffect, useState } from "react";
+import { Link } from "react-router-dom";
+import Menu from "../Menu/Menu";
 import SearchBar from "../SearchBar/SearchBar";
 import './style.css';
 
-const navItems = [
-    { text: 'Home', link: '/' },
-    { text: 'Series', link: '/series' },
-    { text: 'Movies', link: '/movies' },
-    { text: 'New and Popular', link: '/new-popular' },
-    { text: 'My List', link: '/my-list' },
-];
+
 
 const Navbar = () => {
+    const [state, setState] = useState({
+        mobileView: false,
+    });
+
+    const { mobileView } = state;
+
+    useEffect(() => {
+        const setResponsiveness = () => {
+            return window.innerWidth < 900
+                ? setState((prevState) => ({ ...prevState, mobileView: true }))
+                : setState((prevState) => ({ ...prevState, mobileView: false }));
+        };
+
+        setResponsiveness();
+        window.addEventListener("resize", () => setResponsiveness());
+
+        return () => {
+            window.removeEventListener("resize", () => setResponsiveness());
+        }
+    }, []);
 
     return (
         <div sx={{flexGrow: 1}}>
             <AppBar 
                 position="fixed"
-                sx={{display: 'flex', padding: '10px 0'}}
+                sx={{display: 'flex', padding: '10px 0', width: '100%'}}
             >
                 <Toolbar className="nav-toolbar">
                     <div style={{width: '100%'}}>
@@ -30,19 +46,8 @@ const Navbar = () => {
                                     Home
                                 </Link>
                             </div>
-                            <div>
-                                {navItems.map((item) => (
-                                    <NavLink
-                                        to={item.link}
-                                        key={item.text}
-                                        className={({ isActive }) => isActive ? 'nav nav-active' : 'nav nav-inactive'}
-                                        color="info"
-                                        underline="none"
-                                    >
-                                        {item.text}
-                                    </NavLink>
-                                ))}
-                            </div>
+                            {mobileView ? '' : <Menu />}
+                            
                             <div sx={{ display: 'flex', flexWrap: 'wrap', }}>
                                 <SearchBar/>
                                 <Link to={'/login'} className="nav">Login</Link>
